+++ internal/domain/achievement/repository.go
@@ -9,10 +9,10 @@ type Repository interface {
 	// --- 用户成就 ---
 	
 	// FindByUserID 获取用户所有成就
-	FindByUserID(ctx context.Context, userID int64) ([]*UserAchievement, error)
+	FindByUserID(ctx context.Context, userID int) ([]*UserAchievement, error)
 
 	// HasAchievement 检查用户是否已获得某成就
-	HasAchievement(ctx context.Context, userID int64, achievementID int) (bool, error)
+	HasAchievement(ctx context.Context, userID int, achievementID int) (bool, error)
 
 	// Save 保存用户成就
 	Save(ctx context.Context, achievement *UserAchievement) error
