+++ internal/domain/pet/event.go
@@ -6,8 +6,8 @@ import "time"
 
 // PetFedEvent 宠物被喂食事件
 type PetFedEvent struct {
-	PetID     int64     `json:"pet_id"`
-	UserID    int64     `json:"user_id"`
+	PetID     int       `json:"pet_id"`
+	UserID    int       `json:"user_id"`
 	FoodType  int       `json:"food_type"`
 	ExpGained int       `json:"exp_gained"`
 	Timestamp time.Time `json:"timestamp"`
@@ -17,8 +17,8 @@ func (e PetFedEvent) EventName() string { return "pet.fed" }
 
 // PetLevelUpEvent 宠物升级事件
 type PetLevelUpEvent struct {
-	PetID     int64     `json:"pet_id"`
-	UserID    int64     `json:"user_id"`
+	PetID     int       `json:"pet_id"`
+	UserID    int       `json:"user_id"`
 	NewLevel  int       `json:"new_level"`
 	Timestamp time.Time `json:"timestamp"`
 }
@@ -27,8 +27,8 @@ func (e PetLevelUpEvent) EventName() string { return "pet.level_up" }
 
 // PetEvolvedEvent 宠物进化事件
 type PetEvolvedEvent struct {
-	PetID     int64     `json:"pet_id"`
-	UserID    int64     `json:"user_id"`
+	PetID     int       `json:"pet_id"`
+	UserID    int       `json:"user_id"`
 	NewStage  int       `json:"new_stage"`
 	Timestamp time.Time `json:"timestamp"`
 }
@@ -37,8 +37,8 @@ func (e PetEvolvedEvent) EventName() string { return "pet.evolved" }
 
 // PetStatusWarningEvent 宠物状态警告事件
 type PetStatusWarningEvent struct {
-	PetID       int64     `json:"pet_id"`
-	UserID      int64     `json:"user_id"`
+	PetID       int       `json:"pet_id"`
+	UserID      int       `json:"user_id"`
 	WarningType string    `json:"warning_type"` // hungry, unhappy, dirty
 	Timestamp   time.Time `json:"timestamp"`
 }
@@ -47,8 +47,8 @@ func (e PetStatusWarningEvent) EventName() string { return "pet.status_warning" 
 
 // PetCreatedEvent 宠物创建事件
 type PetCreatedEvent struct {
-	PetID     int64     `json:"pet_id"`
-	UserID    int64     `json:"user_id"`
+	PetID     int       `json:"pet_id"`
+	UserID    int       `json:"user_id"`
 	Name      string    `json:"name"`
 	GeneCode  string    `json:"gene_code"`
 	Timestamp time.Time `json:"timestamp"`
