+++ internal/domain/social/gift.go
@@ -6,9 +6,9 @@ import "time"
 
 // GiftRecord 礼物记录实体
 type GiftRecord struct {
-	ID         int64
-	FromUserID int64  // 送礼者
-	ToUserID   int64  // 收礼者
+	ID         int
+	FromUserID int    // 送礼者
+	ToUserID   int    // 收礼者
 	ItemID     int    // 道具ID
 	Quantity   int    // 数量
 	Message    string // 留言
@@ -17,7 +17,7 @@ type GiftRecord struct {
 }
 
 // NewGiftRecord 创建礼物记录
-func NewGiftRecord(fromUserID, toUserID int64, itemID, quantity int, message string) *GiftRecord {
+func NewGiftRecord(fromUserID, toUserID int, itemID, quantity int, message string) *GiftRecord {
 	return &GiftRecord{
 		FromUserID: fromUserID,
 		ToUserID:   toUserID,
@@ -36,9 +36,9 @@ func (g *GiftRecord) MarkAsRead() {
 
 // GiftSentEvent 礼物发送事件
 type GiftSentEvent struct {
-	GiftID     int64     `json:"gift_id"`
-	FromUserID int64     `json:"from_user_id"`
-	ToUserID   int64     `json:"to_user_id"`
+	GiftID     int       `json:"gift_id"`
+	FromUserID int       `json:"from_user_id"`
+	ToUserID   int       `json:"to_user_id"`
 	ItemID     int       `json:"item_id"`
 	Quantity   int       `json:"quantity"`
 	Timestamp  time.Time `json:"timestamp"`
